Move backends.py to avoid circular import
